requirements.txt
certifi==2018.10.15
Django==2.1.4
django-columns==0.1.0
django-crispy-forms==1.7.2
pytz==2018.7
